+++ app/src/main/java/com/google/android/stardroid/layers/IssLayer.kt
@@ -21,8 +21,8 @@ import com.google.android.stardroid.base.TimeConstants
 import com.google.android.stardroid.control.AstronomerModel
 import com.google.android.stardroid.ephemeris.OrbitalElements
 import com.google.android.stardroid.math.Vector3
+import com.google.android.stardroid.renderables.*
 import com.google.android.stardroid.renderer.RendererObjectManager.UpdateType
-import com.google.android.stardroid.source.*
 import com.google.android.stardroid.util.MiscUtil
 import com.google.common.io.Closeables
 import java.io.BufferedReader
