+++ app/src/main/java/com/google/android/stardroid/layers/Layer.kt
@@ -15,7 +15,7 @@ package com.google.android.stardroid.layers
 
 import com.google.android.stardroid.renderer.RendererController
 import com.google.android.stardroid.search.SearchResult
-import com.google.android.stardroid.source.AstronomicalRenderable
+import com.google.android.stardroid.renderables.AstronomicalRenderable
 
 /**
  * A logical collection of objects which should be displayed in SkyMap. For
