+++ app/src/main/java/com/google/android/stardroid/renderables/AstronomicalRenderable.kt
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package com.google.android.stardroid.source
+package com.google.android.stardroid.renderables
 
 import com.google.android.stardroid.math.Vector3
 import com.google.android.stardroid.renderer.RendererObjectManager.UpdateType
