+++ app/src/main/java/com/google/android/stardroid/renderables/LinePrimitive.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.android.stardroid.source;
+package com.google.android.stardroid.renderables;
 
 import java.util.ArrayList;
 import java.util.Collections;
