+++ app/src/main/java/com/google/android/stardroid/renderables/TextPrimitive.java
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 
-package com.google.android.stardroid.source;
+package com.google.android.stardroid.renderables;
 
 import com.google.android.stardroid.math.CoordinateManipulationsKt;
 import com.google.android.stardroid.math.Vector3;
