+++ app/src/main/java/com/google/android/stardroid/renderer/ImageObjectManager.java
@@ -21,7 +21,7 @@ import com.google.android.stardroid.renderer.util.TextureManager;
 import com.google.android.stardroid.renderer.util.TextureReference;
 import com.google.android.stardroid.renderer.util.VertexBuffer;
 import com.google.android.stardroid.math.Vector3;
-import com.google.android.stardroid.source.ImagePrimitive;
+import com.google.android.stardroid.renderables.ImagePrimitive;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
