+++ app/src/main/java/com/google/android/stardroid/views/source/ViewSourceListener.java
@@ -16,10 +16,10 @@ package com.google.android.stardroid.views.source;
 
 import java.util.List;
 
-import com.google.android.stardroid.source.ImagePrimitive;
-import com.google.android.stardroid.source.LinePrimitive;
-import com.google.android.stardroid.source.PointPrimitive;
-import com.google.android.stardroid.source.TextPrimitive;
+import com.google.android.stardroid.renderables.ImagePrimitive;
+import com.google.android.stardroid.renderables.LinePrimitive;
+import com.google.android.stardroid.renderables.PointPrimitive;
+import com.google.android.stardroid.renderables.TextPrimitive;
 
 /**
  * Defines a simple listener interface which responds to changes in the sources and
